Bug 585902: Updated all elements based on facebook.xsd
There were still many places where UIDs and other longs were represented
internally is int.  This is now fixed in all the places I could find.

Bumping FacebookExport addin version

+++ extensions/Exporters/FacebookExport/Mono.Facebook/FriendInfo.cs
@@ -32,10 +32,10 @@ namespace Mono.Facebook
 	public class FriendInfo
 	{
 		[XmlElement ("uid1")]
-		public int UId1;
+		public long UId1;
 
 		[XmlElement ("uid2")]
-		public int UId2;
+		public long UId2;
 
 		[XmlElement ("are_friends")]
 		public bool AreFriends;
